both BP map tests in one file
